fix: white bar on right side of popup on Windows
The white bar thing appears to be a Chromium bug, it is related
to the scroll bar being present on initial render
and then going away. But after it goes away, the width
of the popup remains unchanged for some reason.

This change ensures that there is no scrollbar on initial render
by reducing the amount of initially rendered content.

Somehow the recent "bundle fonts" change surfaced this bug.

Settings "hide" by default is OK because `setPage`
will set class names to proper values anyway
This change simply hides the elements on initial render,
before the scripts are executed