improvement: add CSP & bundle fonts instead of loading from Google (#4)
* improvement: bundle fonts

- improves privacy & security as now nithing from the internet
  is used
- improves font loading reliability when internet is not available

* improvement: add CSP

I looked at the source code and found no external dependencies,
so a strict CSP can be used.

Tested. Works.
- Color picking
- Download image
- Download CSV
- External links

* fix: white bar on right side of popup on Windows

The white bar thing appears to be a Chromium bug, it is related
to the scroll bar being present on initial render
and then going away. But after it goes away, the width
of the popup remains unchanged for some reason.

This change ensures that there is no scrollbar on initial render
by reducing the amount of initially rendered content.

Somehow the recent "bundle fonts" change surfaced this bug.

Settings "hide" by default is OK because `setPage`
will set class names to proper values anyway
This change simply hides the elements on initial render,
before the scripts are executed